requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-Levenshtein
httpx
python-multipart
